RHI improvements (#299)
* GPUBufferDesc.ByteWidth is uint64_t, we don't support D3D11 anymore.
AllocateGPU, UpdateBuffer uses uint64_t.
Rename SubresourceData members with better name.
PipelineStateDesc support custom patch control points, default to 3.

* CreateBuffer accepts const void* instead of const SubresourceData.
Cleanup and improve documentation about buffer creations.

* Rename StructureByteStride to Stride.

* GPUBufferDesc rename ByteWidth to Size.

* CreateBuffer: Under D3D12 - USAGE_READBACK use D3D12_RESOURCE_FLAG_DENY_SHADER_RESOURCE flag and honor BIND_SHADER_RESOURCE.

+++ WickedEngine/wiFont.cpp
@@ -586,7 +586,7 @@ void Draw_internal(const T* text, size_t text_length, const wiFontParams& params
 
 		PushConstantsFont push;
 		push.buffer_index = device->GetDescriptorIndex(&mem.buffer, SRV);
-		push.buffer_offset = mem.offset;
+		push.buffer_offset = (uint32_t)mem.offset;
 		push.texture_index = device->GetDescriptorIndex(&texture, SRV);
 
 		const wiCanvas& canvas = canvases[cmd];
